fix(TodoController): rename namespace
