+++ app/routes/api.php
@@ -3,7 +3,7 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\API\AuthController;
-use App\Http\Controllers\TodoController;
+use App\Http\Controllers\API\TodoController;
 
 Route::get('/user', function (Request $request) {
     return $request->user();
